# sql: remove a bunch of allocations from name and type resolution paths