See individual commits. This PR shaves 3% of the total number of allocations in the FlowSetup benchmark via a variety of methods.

```
name                                          old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=false-24     144µs ± 1%     142µs ± 2%  -1.01%  (p=0.023 n=10+10)

name                                          old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=false-24    28.3kB ± 1%    27.8kB ± 1%  -1.76%  (p=0.000 n=9+9)

name                                          old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=false-24       232 ± 0%       224 ± 0%  -3.45%  (p=0.000 n=8+8)
```