@RaduBerinde 2020-12-18
@jordanlewis @lucy-zhang I have a comment related to the "remove pointless log allocs on the hot path" commit. The very long log line changed by that commit adds the full descriptor information to recorded traces, which makes `SHOW TRACE FOR SESSION` pretty much unusable (see below). Could we backport this commit to 20.2.x and 20.1.x (if the log is there)?

![image](https://user-images.githubusercontent.com/16544120/102663744-4a912e00-4136-11eb-8ae8-59045e42aabd.png)