@jordanlewis 2020-12-19
Hmm... I'm thinking that we actually didn't mean to remove that from the trace. I think I should have used `ExpensiveLogEnabled(ctx,2)` instead of `V(2)`...