@yuzefovich 2020-12-19
Hm, I made a couple of changes like that too, didn't realize `if log.V(n)` would start hiding the log message from the trace.