@RaduBerinde 2020-12-19
Well, as I said, I think we should reconsider dumping that long info in each trace. Or at least split it into multiple logs.